Add limited type-checking support with mypy
This adds necessary requirements/config for type checking
the `lib/` directory, via (after installing the new dependencies).

```
mypy lib/
```

All third-party imports missing declared types are silenced in the
config file, and an extra annotation was added to
lib/elements/validation.py

This is in preparation for new annotations I will add to the
lib/stitches directory to make it a bit clearer what the various types
are.

The typechecking can be run as tests using pytest

```
pytest --mypy
```

Or just with `pytest`, since there are no other tests discovered as of
this commit.